Module 3, Lab 1 Modified

+++ Module03/Lab1/weather.c
@@ -5,16 +5,16 @@
 // ITSC 2181 - Spring 2024
 
 int main() {
-    float temp = 0.0;
+    int temp = 0.0;
 
     int maxDay = 0;
-    float maxTemp = 0.0;
+    int maxTemp = 0.0;
 
     printf ("You will be asked to enter the daily high temperature for 10 consecutive days.\n");
 
     for (int i = 0; i < 10; i++) {
         printf("Enter a high temperature: ");
-        scanf("%f", &temp);
+        scanf("%d", &temp);
 
         if (temp > maxTemp) {
             maxTemp = temp;
@@ -22,6 +22,6 @@ int main() {
         }
     }
 
-    printf("The highest recorded temperature in the 10-day period was: %f degrees", maxTemp);
+    printf("The highest recorded temperature in the 10-day period was: %d degrees", maxTemp);
     printf("\nLast recorded on: Day %d", maxDay);
 }
